# web/src/types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Rheoson API",
    "version": "2.11.0"
  },
  "paths": {
...
    "/api/tracks/{track_id}": {
      "get": {
        "tags": [
          "tracks"
        ],
        "summary": "Get Track",
        "operationId": "get_track_api_tracks__track_id__get",
        "parameters": [
          {
            "name": "track_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Track Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tracks/{track_id}/play": {
      "post": {
        "tags": [
          "tracks"
        ],
        "summary": "Record Play",
        "operationId": "record_play_api_tracks__track_id__play_post",
        "parameters": [
          {
            "name": "track_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Track Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/playlists/": {
      "get": {
        "tags": [
          "playlists"
        ],
        "summary": "List Playlists",
        "description": "List all playlists for the current user.",
        "operationId": "list_playlists_api_playlists__get",
        "parameters": [
          {
            "name": "user",
            "in": "query",
            "required": false,
            "schema": {
              "title": "User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlaylistSchema"
                  },
                  "title": "Response List Playlists Api Playlists  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "playlists"
        ],
        "summary": "Create Playlist",
        "description": "Create a new playlist.",
        "operationId": "create_playlist_api_playlists__post",
        "parameters": [
          {
            "name": "user",
            "in": "query",
            "required": false,
            "schema": {
              "title": "User"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlaylistSchema"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/playlists/{playlist_id}": {
      "get": {
        "tags": [
          "playlists"
        ],
        "summary": "Get Playlist",
        "description": "Get a single playlist by ID.",
        "operationId": "get_playlist_api_playlists__playlist_id__get",
        "parameters": [
          {
            "name": "playlist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Playlist Id"
            }
          },
          {
            "name": "user",
            "in": "query",
            "required": false,
            "schema": {
              "title": "User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "playlists"
        ],
        "summary": "Update Playlist",
        "description": "Update playlist title/description/artwork.",
        "operationId": "update_playlist_api_playlists__playlist_id__patch",
        "parameters": [
          {
            "name": "playlist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Playlist Id"
            }
          },
          {
            "name": "user",
            "in": "query",
            "required": false,
            "schema": {
              "title": "User"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlaylistSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "playlists"
        ],
        "summary": "Delete Playlist",
        "description": "Delete a playlist.\n\nIdempotent: deleting a playlist that no longer exists is a success (204).\nThe frontend queues deletions for offline sync and replays them later \u2014 a\n404 on replay would surface as a spurious sync error.",
        "operationId": "delete_playlist_api_playlists__playlist_id__delete",
        "parameters": [
          {
            "name": "playlist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Playlist Id"
            }
          },
          {
            "name": "user",
            "in": "query",
            "required": false,
            "schema": {
              "title": "User"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/playlists/{playlist_id}/tracks": {
      "post": {
        "tags": [
          "playlists"
        ],
        "summary": "Add Track",
        "description": "Add a track to a playlist.",
        "operationId": "add_track_api_playlists__playlist_id__tracks_post",
        "parameters": [
          {
            "name": "playlist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Playlist Id"
            }
          },
          {
            "name": "user",
            "in": "query",
            "required": false,
            "schema": {
              "title": "User"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/playlists/{playlist_id}/tracks/{track_id}": {
      "delete": {
        "tags": [
          "playlists"
        ],
        "summary": "Remove Track",
        "description": "Remove a track from a playlist.",
        "operationId": "remove_track_api_playlists__playlist_id__tracks__track_id__delete",
        "parameters": [
          {
            "name": "playlist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Playlist Id"
            }
          },
          {
            "name": "track_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Track Id"
            }
          },
          {
            "name": "user",
            "in": "query",
            "required": false,
            "schema": {
              "title": "User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/playlists/{playlist_id}/tracks/reorder": {
      "put": {
        "tags": [
          "playlists"
        ],
        "summary": "Reorder Tracks",
        "description": "Reorder tracks in a playlist.",
        "operationId": "reorder_tracks_api_playlists__playlist_id__tracks_reorder_put",
        "parameters": [
          {
            "name": "playlist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Playlist Id"
            }
          },
          {
            "name": "user",
            "in": "query",
            "required": false,
            "schema": {
              "title": "User"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/playlists/{playlist_id}/import": {
      "post": {
        "tags": [
          "playlists"
        ],
        "summary": "Import Spotify",
        "description": "Import a Spotify playlist into a local playlist.",
        "operationId": "import_spotify_api_playlists__playlist_id__import_post",
        "parameters": [
          {
            "name": "playlist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Playlist Id"
            }
          },
          {
            "name": "user",
            "in": "query",
            "required": false,
            "schema": {
              "title": "User"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/playlists/import": {
      "post": {
        "tags": [
          "playlists"
        ],
        "summary": "Import Playlist Url",
        "description": "Create a new playlist from a shared URL (Spotify/YouTube/SoundCloud\u2026).\n\nResolves the URL server-side and stores the resolved track IDs in a\nbrand-new playlist. This backs the Library \"Import playlist\" flow.",
        "operationId": "import_playlist_url_api_playlists_import_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportPlaylistRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/playlists/{playlist_id}/export": {
      "get": {
        "tags": [
          "playlists"
        ],
        "summary": "Export Playlist",
        "description": "Export playlist as JSON with hydrated track metadata.",
        "operationId": "export_playlist_api_playlists__playlist_id__export_get",
        "parameters": [
          {
            "name": "playlist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Playlist Id"
            }
          },
          {
            "name": "user",
            "in": "query",
            "required": false,
            "schema": {
              "title": "User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TrackSchema": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "default": ""
          },
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          },
          "artist": {
            "$ref": "#/components/schemas/ArtistSchema"
          },
          "album": {
            "$ref": "#/components/schemas/AlbumSchema"
          },
          "artworkUrl": {
            "type": "string",
            "title": "Artworkurl",
            "default": ""
          },
          "duration": {
            "type": "number",
            "title": "Duration",
            "default": 0.0
          },
          "streamUrl": {
            "type": "string",
            "title": "Streamurl",
            "default": ""
          },
          "filePath": {
            "type": "string",
            "title": "Filepath",
            "default": ""
          },
          "isDownloaded": {
            "type": "boolean",
            "title": "Isdownloaded",
            "default": false
          },
          "isLiked": {
            "type": "boolean",
            "title": "Isliked",
            "default": false
          },
          "youtubeId": {
            "type": "string",
            "title": "Youtubeid",
            "default": ""
          },
          "spotifyId": {
            "type": "string",
            "title": "Spotifyid",
            "default": ""
          },
          "addedAt": {
            "type": "string",
            "title": "Addedat",
            "default": ""
          },
          "trackNumber": {
            "type": "integer",
            "title": "Tracknumber",
            "default": 0
          },
          "playCount": {
            "type": "integer",
            "title": "Playcount",
            "default": 0
          }
        },
        "type": "object",
        "title": "TrackSchema",
        "required": [
          "addedAt",
          "album",
          "artist",
          "filePath",
          "playCount",
          "spotifyId",
          "streamUrl",
          "trackNumber",
          "youtubeId"
        ]
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}